AG-15592 Move `CollapseWidgetEvent` and `ExpandWidgetEvent`
